# gonna-pay-backend/docs/swagger.yaml
basePath: /
definitions:
...
  request.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
    required:
    - email
    - name
    - password
    - phone
    type: object
...
  request.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
    type: object
...
  response.PaginatedUserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.UserResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  response.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  rest_errors.Causes:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  rest_errors.RestErrors:
    properties:
      causes:
        items:
          $ref: '#/definitions/rest_errors.Causes'
        type: array
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
host: localhost:3333
info:
  contact: {}
  description: API de gestão financeira e divisão de custos
  title: Gonna Pay API
  version: "1.0"
paths:
...
  /user:
    post:
      consumes:
      - application/json
      description: Cria um novo usuário (registro). Não requer autenticação
      parameters:
      - description: Dados do usuário
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "409":
          description: Email já cadastrado
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      summary: Criar usuário
      tags:
      - users
  /user/{id}:
    delete:
      description: Remove um usuário pelo ID
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Deletar usuário
      tags:
      - users
    get:
      description: Retorna um usuário pelo ID
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Buscar usuário por ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um usuário pelo ID
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      - description: Dados para atualização
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "409":
          description: Email já em uso
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Atualizar usuário
      tags:
      - users
  /users:
    get:
      description: Retorna todos os usuários cadastrados
      parameters:
      - description: 'Número da página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Itens por página (padrão: 20, máximo: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginatedUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Listar usuários
      tags:
      - users
schemes:
- http
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"